requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==25.4.0
babel==2.18.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
daff==1.4.2
dbt-adapters==1.17.3
dbt-common==1.34.1
dbt-core==1.10.19
dbt-duckdb==1.10.0
dbt-extractor==0.6.0
dbt-protos==1.0.431
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.1
duckdb==1.4.4
idna==3.11
importlib_metadata==8.7.1
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
leather==0.4.1
MarkupSafe==3.0.3
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.2.1
numpy==2.0.2
orderly-set==5.5.0
packaging==26.0
pandas==2.3.3
parsedatetime==2.6
pathspec==0.12.1
protobuf==6.33.5
pyarrow==21.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
snowplow-tracker==1.1.0
snowflake-connector-python==4.3.0
sqlparse==0.5.4
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
zipp==3.23.0
